backend/pyproject.toml
[project]
name = "django-react-todo"
version = "0.1.0"
description = "Todo アプリケーションのバックエンド"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.4",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "autopep8>=2.3.1",
    "email-validator>=2.3.0",
]

[dependency-groups]
dev = [
    "black>=24.10.0",
    "inflection>=0.5.1",
    "pyyaml>=6.0.3",
    "uritemplate>=4.2.0",
]

[tool.uv.sources]
